requirements.txt
discord.py==2.1.1
requests==2.28.1
